+++ zcash_client_backend/src/wallet.rs
@@ -15,17 +15,66 @@ use zcash_primitives::{
         },
         TxId,
     },
-    zip32::AccountId,
+    zip32::{AccountId, Scope},
 };
+
+use crate::{address::UnifiedAddress, PoolType, ShieldedProtocol};
+
+/// A unique identifier for a shielded transaction output
+#[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord)]
+pub struct NoteId {
+    txid: TxId,
+    protocol: ShieldedProtocol,
+    output_index: u16,
+}
+
+impl NoteId {
+    /// Constructs a new `NoteId` from its parts.
+    pub fn new(txid: TxId, protocol: ShieldedProtocol, output_index: u16) -> Self {
+        Self {
+            txid,
+            protocol,
+            output_index,
+        }
+    }
+
+    /// Returns the ID of the transaction containing this note.
+    pub fn txid(&self) -> &TxId {
+        &self.txid
+    }
+
+    /// Returns the shielded protocol used by this note.
+    pub fn protocol(&self) -> ShieldedProtocol {
+        self.protocol
+    }
+
+    /// Returns the index of this note within its transaction's corresponding list of
+    /// shielded outputs.
+    pub fn output_index(&self) -> u16 {
+        self.output_index
+    }
+}
+
+/// A type that represents the recipient of a transaction output: a recipient address (and, for
+/// unified addresses, the pool to which the payment is sent) in the case of an outgoing output, or an
+/// internal account ID and the pool to which funds were sent in the case of a wallet-internal
+/// output.
+#[derive(Debug, Clone)]
+pub enum Recipient {
+    Transparent(TransparentAddress),
+    Sapling(sapling::PaymentAddress),
+    Unified(UnifiedAddress, PoolType),
+    InternalAccount(AccountId, PoolType),
+}
 
 /// A subset of a [`Transaction`] relevant to wallets and light clients.
 ///
 /// [`Transaction`]: zcash_primitives::transaction::Transaction
-pub struct WalletTx<N> {
+pub struct WalletTx<N, S> {
     pub txid: TxId,
     pub index: usize,
     pub sapling_spends: Vec<WalletSaplingSpend>,
-    pub sapling_outputs: Vec<WalletSaplingOutput<N>>,
+    pub sapling_outputs: Vec<WalletSaplingOutput<N, S>>,
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -109,8 +158,17 @@ impl WalletSaplingSpend {
 
 /// A subset of an [`OutputDescription`] relevant to wallets and light clients.
 ///
+/// The type parameter `<N>` is used to specify the nullifier type, which may vary between
+/// `Sapling` and `Orchard`, and also may vary depending upon the type of key that was used to
+/// decrypt this output; incoming viewing keys do not have the capability to derive the nullifier
+/// for a note, and the `<N>` will be `()` in these cases.
+///
+/// The type parameter `<S>` is used to specify the type of the scope of the key used to recover
+/// this output; this will usually be [`zcash_primitives::zip32::Scope`] for received notes, and
+/// `()` for sent notes.
+///
 /// [`OutputDescription`]: zcash_primitives::transaction::components::OutputDescription
-pub struct WalletSaplingOutput<N> {
+pub struct WalletSaplingOutput<N, S> {
     index: usize,
     cmu: sapling::note::ExtractedNoteCommitment,
     ephemeral_key: EphemeralKeyBytes,
@@ -119,9 +177,10 @@ pub struct WalletSaplingOutput<N> {
     is_change: bool,
     note_commitment_tree_position: Position,
     nf: N,
+    recipient_key_scope: S,
 }
 
-impl<N> WalletSaplingOutput<N> {
+impl<N, S> WalletSaplingOutput<N, S> {
     /// Constructs a new `WalletSaplingOutput` value from its constituent parts.
     #[allow(clippy::too_many_arguments)]
     pub fn from_parts(
@@ -133,6 +192,7 @@ impl<N> WalletSaplingOutput<N> {
         is_change: bool,
         note_commitment_tree_position: Position,
         nf: N,
+        recipient_key_scope: S,
     ) -> Self {
         Self {
             index,
@@ -143,6 +203,7 @@ impl<N> WalletSaplingOutput<N> {
             is_change,
             note_commitment_tree_position,
             nf,
+            recipient_key_scope,
         }
     }
 
@@ -170,38 +231,58 @@ impl<N> WalletSaplingOutput<N> {
     pub fn nf(&self) -> &N {
         &self.nf
     }
+    pub fn recipient_key_scope(&self) -> &S {
+        &self.recipient_key_scope
+    }
+}
+
+/// An enumeration of supported shielded note types for use in [`ReceivedNote`]
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum WalletNote {
+    Sapling(sapling::Note),
+    Orchard(orchard::Note),
+}
+
+impl WalletNote {
+    pub fn value(&self) -> NonNegativeAmount {
+        match self {
+            WalletNote::Sapling(n) => n.value().try_into().expect(
+                "Sapling notes must have values in the range of valid non-negative ZEC values.",
+            ),
+            WalletNote::Orchard(n) => NonNegativeAmount::from_u64(n.value().inner()).expect(
+                "Orchard notes must have values in the range of valid non-negative ZEC values.",
+            ),
+        }
+    }
 }
 
 /// Information about a note that is tracked by the wallet that is available for spending,
 /// with sufficient information for use in note selection.
 #[derive(Debug, Clone, PartialEq, Eq)]
-pub struct ReceivedSaplingNote<NoteRef> {
+pub struct ReceivedNote<NoteRef> {
     note_id: NoteRef,
     txid: TxId,
     output_index: u16,
-    diversifier: sapling::Diversifier,
-    note_value: NonNegativeAmount,
-    rseed: sapling::Rseed,
+    note: WalletNote,
+    spending_key_scope: Scope,
     note_commitment_tree_position: Position,
 }
 
-impl<NoteRef> ReceivedSaplingNote<NoteRef> {
+impl<NoteRef> ReceivedNote<NoteRef> {
     pub fn from_parts(
         note_id: NoteRef,
         txid: TxId,
         output_index: u16,
-        diversifier: sapling::Diversifier,
-        note_value: NonNegativeAmount,
-        rseed: sapling::Rseed,
+        note: WalletNote,
+        spending_key_scope: Scope,
         note_commitment_tree_position: Position,
     ) -> Self {
-        ReceivedSaplingNote {
+        ReceivedNote {
             note_id,
             txid,
             output_index,
-            diversifier,
-            note_value,
-            rseed,
+            note,
+            spending_key_scope,
             note_commitment_tree_position,
         }
     }
@@ -209,34 +290,33 @@ impl<NoteRef> ReceivedSaplingNote<NoteRef> {
     pub fn internal_note_id(&self) -> &NoteRef {
         &self.note_id
     }
-
     pub fn txid(&self) -> &TxId {
         &self.txid
     }
     pub fn output_index(&self) -> u16 {
         self.output_index
     }
-    pub fn diversifier(&self) -> sapling::Diversifier {
-        self.diversifier
+    pub fn note(&self) -> &WalletNote {
+        &self.note
     }
     pub fn value(&self) -> NonNegativeAmount {
-        self.note_value
+        self.note.value()
     }
-    pub fn rseed(&self) -> sapling::Rseed {
-        self.rseed
+    pub fn spending_key_scope(&self) -> Scope {
+        self.spending_key_scope
     }
     pub fn note_commitment_tree_position(&self) -> Position {
         self.note_commitment_tree_position
     }
 }
 
-impl<NoteRef> sapling_fees::InputView<NoteRef> for ReceivedSaplingNote<NoteRef> {
+impl<NoteRef> sapling_fees::InputView<NoteRef> for ReceivedNote<NoteRef> {
     fn note_id(&self) -> &NoteRef {
         &self.note_id
     }
 
     fn value(&self) -> NonNegativeAmount {
-        self.note_value
+        self.note.value()
     }
 }