+++ zcash_client_sqlite/src/testing.rs
@@ -708,7 +708,7 @@ impl<Cache> TestState<Cache> {
         min_confirmations: u32,
     ) -> NonNegativeAmount {
         self.with_account_balance(account, min_confirmations, |balance| {
-            balance.sapling_balance.spendable_value
+            balance.sapling_balance().spendable_value()
         })
     }
 
@@ -718,8 +718,8 @@ impl<Cache> TestState<Cache> {
         min_confirmations: u32,
     ) -> NonNegativeAmount {
         self.with_account_balance(account, min_confirmations, |balance| {
-            balance.sapling_balance.value_pending_spendability
-                + balance.sapling_balance.change_pending_confirmation
+            balance.sapling_balance().value_pending_spendability()
+                + balance.sapling_balance().change_pending_confirmation()
         })
         .unwrap()
     }
@@ -731,7 +731,7 @@ impl<Cache> TestState<Cache> {
         min_confirmations: u32,
     ) -> NonNegativeAmount {
         self.with_account_balance(account, min_confirmations, |balance| {
-            balance.sapling_balance.change_pending_confirmation
+            balance.sapling_balance().change_pending_confirmation()
         })
     }